unarchiving/requirements.txt
boto3==1.8.3
pyobjc==4.2.2
unidecode==1.0.22
